+++ debug_toolbar/panels/sql/utils.py
@@ -86,6 +86,11 @@ class EscapedStringSerializer:
         return "".join(escaped_value(token) for token in stmt.flatten())
 
 
+def is_select_query(sql):
+    # UNION queries can start with "(".
+    return sql.lower().lstrip(" (").startswith("select")
+
+
 def reformat_sql(sql, *, with_toggle=False):
     formatted = parse_sql(sql)
     if not with_toggle: