+++ docs/changes.rst
@@ -3,6 +3,7 @@ Change log
 
 Pending
 -------
+* Support select and explain buttons for ``UNION`` queries on PostgreSQL.
 
 * Fixed internal toolbar requests being instrumented if the Django setting
   ``FORCE_SCRIPT_NAME`` was set.